fix: use select loop to ensure all passwords are attempted

+++ internal/worker/state.go
@@ -9,7 +9,7 @@ type WorkerState struct {
 	ChunkID       int    `json:"chunk_id"`
 	PasswordIndex int    `json:"password_index"`
 
-	CompeletedPasswords map[int]bool `json:"completed_passwords"`
+	CompeletedPasswords map[uint64]bool `json:"completed_passwords"`
 
 	Payload shared.PayloadJobDetails `json:"payload"`
 }
@@ -33,7 +33,7 @@ func InitialState() WorkerState {
 		ID:                  "",
 		ChunkID:             0,
 		PasswordIndex:       0,
-		CompeletedPasswords: map[int]bool{},
+		CompeletedPasswords: map[uint64]bool{},
 		Payload:             shared.PayloadJobDetails{},
 	}
 }
@@ -43,7 +43,7 @@ func InitialStateWithID(id string) WorkerState {
 		ID:                  id,
 		ChunkID:             0,
 		PasswordIndex:       0,
-		CompeletedPasswords: map[int]bool{},
+		CompeletedPasswords: map[uint64]bool{},
 		Payload:             shared.PayloadJobDetails{},
 	}
 }
